refactor: 웹뷰를 컴포넌트 단위로 응집하고 웹 상태머신 제거
- 과도한 추상화 레이어 제거: WebView, View 제거
- 응집도를 고려해 컴포넌트 단위로 뷰 재구성: PurchaseView, WinningNumbersView
- 상태머신으로 인한 코드 복잡성 제거
